+++ T2/system/system.c
@@ -2,16 +2,16 @@
 
 Interval_t *retrossubs(IntervalMatrix_t *A)
 {
-    int n = A->rows;
-    int m = A->cols;
+    lli n = A->rows;
+    lli m = A->cols;
     Interval_t *x = malloc(sizeof(Interval_t) * n);
 
     // for each row
-    for (int i = n - 1; i >= 0; i--)
+    for (lli i = n - 1; i >= 0; i--)
     {
         x[i] = A->independent_terms[i];
         // for each column
-        for (int j = i + 1; j < m; j++)
+        for (lli j = i + 1; j < m; j++)
             // calculate the sum of the products
             x[i] = op_sub_interval(x[i], op_mul_interval(A->data[i][j], x[j]));
         // calculate the value of x(the independent term)
@@ -25,25 +25,25 @@ IntervalMatrix_t *partial_pivoting_system_solver(IntervalMatrix_t *A)
 {
     Interval_t multiplier;
 
-    int n = A->rows;
+    lli n = A->rows;
     // for each row
-    for (int i = 0; i < n; i++)
+    for (lli i = 0; i < n; i++)
     {
         // find the pivot
-        int pivot = find_partial_pivot(A, i, i);
+        lli pivot = find_partial_pivot(A, i, i);
         // swap rows if necessary
         if (pivot != i)
             swap_rows(A, i, pivot);
 
         // for each column
-        for (int j = i + 1; j < n; j++)
+        for (lli j = i + 1; j < n; j++)
         {
             // calculate the multiplier
             multiplier = op_div_interval(A->data[j][i], A->data[i][i]);
             A->data[j][i].min.f = 0.0;
             A->data[j][i].max.f = 0.0;
             // for each element in the row
-            for (int k = i + 1; k < n; k++)
+            for (lli k = i + 1; k < n; k++)
             {
                 // calculate the new value
                 A->data[j][k] = op_sub_interval(A->data[j][k], op_mul_interval(A->data[i][k], multiplier));
@@ -58,9 +58,9 @@ IntervalMatrix_t *partial_pivoting_system_solver(IntervalMatrix_t *A)
 
 void print_system(IntervalMatrix_t A)
 {
-    for (int i = 0; i < A.rows; i++)
+    for (lli i = 0; i < A.rows; i++)
     {
-        for (int j = 0; j < A.cols; j++)
+        for (lli j = 0; j < A.cols; j++)
         {
             printf("[%1.8e, %1.8e] ", A.data[i][j].min.f, A.data[i][j].max.f);
         }
@@ -69,16 +69,16 @@ void print_system(IntervalMatrix_t A)
     }
 }
 
-int find_partial_pivot(IntervalMatrix_t *A, int row, int col)
+lli find_partial_pivot(IntervalMatrix_t *A, lli row, lli col)
 {
-    int imax = row;
-    for (int i = row; i < A->rows; i++)
+    lli imax = row;
+    for (lli i = row; i < A->rows; i++)
         if (greater_than(A->data[i][col], A->data[imax][col]))
             imax = i;
     return imax;
 }
 
-void swap_rows(IntervalMatrix_t *A, int row1, int row2)
+void swap_rows(IntervalMatrix_t *A, lli row1, lli row2)
 {
     Interval_t *aux = A->data[row1];
     A->data[row1] = A->data[row2];
@@ -89,17 +89,17 @@ void swap_rows(IntervalMatrix_t *A, int row1, int row2)
     A->independent_terms[row2] = aux2;
 }
 
-Interval_t *show_residual(IntervalMatrix_t *A, Interval_t *solution, IntervalPoint_t *table, int k)
+Interval_t *show_residual(IntervalMatrix_t *A, Interval_t *solution, IntervalPoint_t *table, lli k)
 {
-    int m = A->cols;
+    lli m = A->cols;
     Interval_t *residual = malloc(sizeof(Interval_t) * k);
-    for (int i = 0; i < k; i++)
+    for (lli i = 0; i < k; i++)
     {
         residual[i].max.f = 0.0;
         residual[i].min.f = 0.0;
 
         // for each column
-        for (int j = 0; j < m; j++)
+        for (lli j = 0; j < m; j++)
         {
             residual[i] = op_sum_interval(residual[i], op_mul_interval(solution[j], op_pow_interval(table[i].x, j)));
         }
